requirements.txt
blinker==1.9.0
click==8.2.1
Deprecated==1.2.18
Flask==3.1.1
flask-cors==6.0.1
Flask-Limiter==3.12
itsdangerous==2.2.0
Jinja2==3.1.6
limits==5.4.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ordered-set==4.1.0
packaging==25.0
Pygments==2.19.2
python-dotenv==1.1.1
redis==6.2.0
rich==13.9.4
typing_extensions==4.14.0
Werkzeug==3.1.3
wrapt==1.17.2
